go/common/node: Move HashRAK function to a better place
